+++ lib/Subscriber/Phpcr/ReorderSubscriber.php
@@ -11,14 +11,10 @@
 
 namespace Sulu\Component\DocumentManager\Subscriber\Phpcr;
 
-use PHPCR\NodeInterface;
-use PHPCR\Util\PathHelper;
-use PHPCR\Util\UUIDHelper;
-use Sulu\Component\DocumentManager\DocumentRegistry;
 use Sulu\Component\DocumentManager\Event\ReorderEvent;
 use Sulu\Component\DocumentManager\Events;
 use Sulu\Component\DocumentManager\Exception\DocumentManagerException;
-use Sulu\Component\DocumentManager\NodeManager;
+use Sulu\Component\DocumentManager\NodeHelperInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -27,23 +23,13 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class ReorderSubscriber implements EventSubscriberInterface
 {
     /**
-     * @var NodeManager
+     * @var NodeHelperInterface
      */
-    private $nodeManager;
+    private $nodeHelper;
 
-    /**
-     * @var DocumentRegistry
-     */
-    private $documentRegistry;
-
-    /**
-     * @param NodeManager $nodeManager
-     * @param DocumentRegistry $documentRegistry
-     */
-    public function __construct(NodeManager $nodeManager, DocumentRegistry $documentRegistry)
+    public function __construct(NodeHelperInterface $nodeHelper)
     {
-        $this->nodeManager = $nodeManager;
-        $this->documentRegistry = $documentRegistry;
+        $this->nodeHelper = $nodeHelper;
     }
 
     /**
@@ -65,72 +51,6 @@ class ReorderSubscriber implements EventSubscriberInterface
      */
     public function handleReorder(ReorderEvent $event)
     {
-        $document = $event->getDocument();
-        $siblingId = $event->getDestId();
-        $after = $event->getAfter();
-
-        $node = $this->documentRegistry->getNodeForDocument($document);
-        $parentNode = $node->getParent();
-
-        $nodeName = $node->getName();
-        $siblingName = $this->resolveSiblingName($siblingId, $parentNode, $node);
-        if (true === $after) {
-            $siblingName = $this->resolveAfterSiblingName($parentNode, $siblingName);
-        }
-
-        $parentNode->orderBefore($nodeName, $siblingName);
-    }
-
-    private function resolveSiblingName($siblingId, NodeInterface $parentNode, NodeInterface $node)
-    {
-        if (null === $siblingId) {
-            return;
-        }
-
-        $siblingPath = $siblingId;
-        if (UUIDHelper::isUUID($siblingId)) {
-            $siblingPath = $this->nodeManager->find($siblingId)->getPath();
-        }
-
-        if ($siblingPath !== null && PathHelper::getParentPath($siblingPath) !== $parentNode->getPath()) {
-            throw new DocumentManagerException(sprintf(
-                'Cannot reorder documents which are not siblings. Trying to reorder "%s" to "%s"',
-                $node->getPath(), $siblingPath
-            ));
-        }
-
-        if (null !== $siblingPath) {
-            return PathHelper::getNodeName($siblingPath);
-        }
-
-        return $node->getName();
-    }
-
-    /**
-     * If the node should be ordered after the target then we need to order the
-     * node before the sibling after the target sibling. If the node should be the
-     * last sibling, then the target sibling should be NULL.
-     *
-     * @param NodeInterface $parentNode
-     * @param string $siblingName
-     *
-     * @return string
-     */
-    private function resolveAfterSiblingName(NodeInterface $parentNode, $siblingName)
-    {
-        $targetName = null;
-        $found = false;
-
-        foreach (array_keys($parentNode->getNodes()) as $name) {
-            if ($name === $siblingName) {
-                $found = true;
-                continue;
-            } elseif ($found) {
-                $targetName = $name;
-                break;
-            }
-        }
-
-        return $targetName;
+        $this->nodeHelper->reorder($event->getNode(), $event->getDestId());
     }
 }